chore: convert test/utils to ts
This patch converts the test utils to TS, and also removes a few lines
of commented-out code. Note that I haven't change the import style, as I
can't seem to make it play nicely with TAP.
